fix again

+++ research/object_detection/meta_architectures/detr_meta_arch.py
@@ -109,7 +109,7 @@ class DETRMetaArch(model.DetectionModel):
 
     fake_logits = np.zeros((91))
     fake_logits[5] = 1
-    logits = tf.convert_to_tensor(fake_logits)
+    logits = tf.convert_to_tensor(fake_logits, dtype=tf.float32)
     #bboxes_encoded = self._bbox_ffn(bboxes_encoded) #tf.keras.backend.sigmoid(bboxes_encoded)
     bboxes_encoded = ops.normalized_to_image_coordinates(
         bboxes_encoded, image_shape, self._parallel_iterations)
